update(debug): add print is_quoted_arr() and sep file for norm
