+++ Makefile
@@ -71,6 +71,7 @@ DEBUG_DIR	:=	debug
 SRCS		+=	$(DEBUG_DIR)/debug_print_ast.c \
 				$(DEBUG_DIR)/debug_print_ast_sub.c \
 				$(DEBUG_DIR)/debug_token_dq.c \
+				$(DEBUG_DIR)/debug_token_dq_sub.c \
 				$(DEBUG_DIR)/put.c
 
 EXEC_DIR	:=	exec
