+++ srcs/debug/debug_token_dq.c
@@ -12,47 +12,25 @@
 //}	t_token;
 //
 
-char	*get_token_kind_str(t_token_kind kind)
-{
-	if (kind == TOKEN_KIND_WORD)
-		return ("w");
-	if (kind == TOKEN_KIND_OP_PIPE)
-		return ("|");
-	if (kind == TOKEN_KIND_OP_OR)
-		return ("||");
-	if (kind == TOKEN_KIND_OP_AND)
-		return ("&&");
-	if (kind == TOKEN_KIND_REDIRECT_IN)
-		return ("<");
-	if (kind == TOKEN_KIND_REDIRECT_HEREDOC)
-		return ("<<");
-	if (kind == TOKEN_KIND_REDIRECT_OUT)
-		return (">");
-	if (kind == TOKEN_KIND_REDIRECT_APPEND)
-		return (">>");
-	if (kind == TOKEN_KIND_PAREN_LEFT)
-		return ("(");
-	if (kind == TOKEN_KIND_PAREN_RIGHT)
-		return (")");
-	return ("Error");
-}
-
-static char	*get_quote_str(t_quote quote)
-{
-	if (quote == QUOTE_SINGLE)
-		return ("\'");
-	if (quote == QUOTE_DOUBLE)
-		return ("\"");
-	if (quote == QUOTE_NONE)
-		return ("");
-	return ("Error");
-}
-
 static char	*get_concat_str(bool is_concat_next)
 {
 	if (is_concat_next)
 		return ("=");
 	return ("");
+}
+
+static void	print_is_quoted_arr(t_token *token)
+{
+	size_t	i;
+
+	ft_dprintf(2, "  is_quoted:[");
+	i = 0;
+	while (token->str && token->str[i])
+	{
+		ft_dprintf(2, "%d", token->is_quoted_arr[i]);
+		i++;
+	}
+	ft_dprintf(2, "]\n");
 }
 
 void	debug_token_dq_node(t_deque_node *node)
@@ -62,10 +40,11 @@ void	debug_token_dq_node(t_deque_node *node)
 	while (node)
 	{
 		token = (t_token *)node->content;
-		ft_dprintf(2, "  str   :[%s]\n", token->str);
-		ft_dprintf(2, "  kind  :[%s]\n", get_token_kind_str(token->kind));
-		ft_dprintf(2, "  quote :[%s]\n", get_quote_str(token->quote));
-		ft_dprintf(2, "  concat:[%s]\n", get_concat_str(token->concat_next));
+		ft_dprintf(2, "  str      :[%s]\n", token->str);
+		print_is_quoted_arr(token);
+		ft_dprintf(2, "  kind     :[%s]\n", get_token_kind_str(token->kind));
+		ft_dprintf(2, "  quote    :[%s]\n", get_quote_str(token->quote));
+		ft_dprintf(2, "  concat   :[%s]\n", get_concat_str(token->concat_next));
 		node = node->next;
 		if (node)
 			ft_dprintf(STDERR_FILENO, "\n");
